asistencia/proyecto_qr/app/requirements.txt
Flask==2.3.0
mysql-connector-python==8.0.31
qrcode[pil]
pytz==2022.1
requests==2.28.1
pyotp==2.7.0
